CM-1632: Extracted artifact support test cases into ArtifactSupportTest.
